refactor models; add separate suggestion type for collection endpoints
